+++ gems/aws-sdk-sqs/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Support VPC Endpoint pattern in Aws::SQS::Plugins::QueueUrl (Github #2114)
+
 1.23.0 (2019-10-23)
 ------------------